split up result_ontology.py
