+++ assemblyline/odm/models/ontology/types/sandbox.py
@@ -1,0 +1,6 @@
+from assemblyline import odm
+
+
+class Sandbox(odm.Model):
+    # Perceived capabilities that the file may have
+    capabilities = odm.Optional(odm.List(odm.Text()))
